refactor(D4): extract useManualSimulation hook from App.tsx
First stage of D4 (relieve the App.tsx hub). Moves the two operator-
driven "simulate now" flows and the interactive SLD-edit state they
share out of App.tsx into hooks/useManualSimulation.ts:

- handleSimulateUnsimulatedAction (double-click an un-simulated overview
  pin), handleSimulateSldEdit (staged breaker/injection edit → action),
  the useSldTopologyEdit state, the implicit edit-mode sync, and the
  debounced target-topology preview.

App owns the collaborators (diagrams, analysis result, the manual-action
wrapper) and passes them into the hook, which returns the SLD-edit state
+ the two handlers the JSX consumes — a thin orchestrator, not a second
state hub. Behaviour-preserving (full Vitest suite green).

App.tsx 2048 → 1795 lines; APP_TSX_MAX ratcheted 2100 → 1850 to lock the
win. New hooks/useManualSimulation.test.ts covers the no-contingency
guard, the metrics+diagram stream (prime + provenance), the stream-error
path, the empty SLD-edit no-op, and sldEditBaseActionId derivation.

Remaining D4 tail (bigger bets, tracked): split useDiagrams into
per-domain hooks behind the DiagramsState facade, and consolidate the
exploded VisualizationPanel / ActionFeed props into cohesive state-object
props.
